+++ library/src/main/java/com/avast/android/dialogs/core/BaseDialogFragment.java
@@ -16,6 +16,10 @@
 
 package com.avast.android.dialogs.core;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
 import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
@@ -23,6 +27,7 @@ import android.graphics.Typeface;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.v4.app.DialogFragment;
+import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
 import android.text.TextUtils;
@@ -32,6 +37,7 @@ import android.view.ViewGroup;
 import android.widget.*;
 
 import com.avast.android.dialogs.R;
+import com.avast.android.dialogs.iface.ISimpleDialogCancelListener;
 import com.avast.android.dialogs.util.TypefaceHelper;
 
 /**
@@ -40,6 +46,8 @@ import com.avast.android.dialogs.util.TypefaceHelper;
  * @author David Vávra (david@inmite.eu)
  */
 public abstract class BaseDialogFragment extends DialogFragment implements DialogInterface.OnShowListener {
+
+    protected int mRequestCode;
 
     @NonNull
     @Override
@@ -61,6 +69,26 @@ public abstract class BaseDialogFragment extends DialogFragment implements Dialo
         return build(builder).create();
     }
 
+    @Override
+    public void onActivityCreated(Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+        final Fragment targetFragment = getTargetFragment();
+        if (targetFragment != null) {
+            mRequestCode = getTargetRequestCode();
+        } else {
+            Bundle args = getArguments();
+            if (args != null) {
+                mRequestCode = args.getInt(BaseDialogBuilder.ARG_REQUEST_CODE, 0);
+            }
+        }
+    }
+
+    /** Key method for using {@link com.avast.android.dialogs.core.BaseDialogFragment}.
+     *  Customized dialogs need to be set up via provided builder.
+     *
+     * @param initialBuilder Provided builder for setting up customized dialog
+     * @return Updated builder
+     */
     protected abstract Builder build(Builder initialBuilder);
 
     @Override
@@ -86,6 +114,43 @@ public abstract class BaseDialogFragment extends DialogFragment implements Dialo
             boolean scrollable = isScrollable(vScrollView) || isScrollable(vListView);
             modifyButtonsBasedOnScrollableContent(scrollable);
         }
+    }
+
+    @Override
+    public void onCancel(DialogInterface dialog) {
+        super.onCancel(dialog);
+        for (ISimpleDialogCancelListener listener : getCancelListeners()) {
+            listener.onCancelled(mRequestCode);
+        }
+    }
+
+    /** Get dialog cancel listeners.
+     *  There might be more than one cancel listener.
+     *
+     * @return Dialog cancel listeners
+     * @since 2.1.0
+     */
+    protected List<ISimpleDialogCancelListener> getCancelListeners() {
+        return getDialogListeners(ISimpleDialogCancelListener.class);
+    }
+
+    /** Utility method for acquiring all listeners of some type for current instance of DialogFragment
+     *
+     * @param listenerInterface Interface of the desired listeners
+     * @return Unmodifiable list of listeners
+     * @since 2.1.0
+     */
+    @SuppressWarnings("unchecked")
+    protected <T> List<T> getDialogListeners(Class<T> listenerInterface) {
+        final Fragment targetFragment = getTargetFragment();
+        List<T> listeners = new ArrayList<T>(2);
+        if (targetFragment != null && listenerInterface.isAssignableFrom(targetFragment.getClass())) {
+            listeners.add((T) targetFragment);
+        }
+        if (getActivity() != null && listenerInterface.isAssignableFrom(getActivity().getClass())) {
+            listeners.add((T) getActivity());
+        }
+        return Collections.unmodifiableList(listeners);
     }
 
     /**